fix: resolve alhai_reports build errors — screenWidth, double→int types
- Replace context.screenWidth/screenHeight with MediaQuery.of(context).size
- Change LowStockItem and _InventoryItem stock fields from int to double
  to match database schema (stockQty/minQty are doubles)

+++ packages/alhai_reports/lib/src/screens/reports/inventory_report_screen.dart
@@ -312,7 +312,7 @@ class _InventoryReportScreenState extends ConsumerState<InventoryReportScreen> {
 
 class _InventoryItem {
   final String name, category, sku;
-  final int stock, minStock;
+  final double stock, minStock;
   final double cost, price;
   final DateTime lastUpdated;
   _InventoryItem(
